Refactor the pub tests.
The goal of this refactoring is to make it easier to flexibly set up the
filesystem before Pub runs and validate it afterwards. To do so, the pub tests
are made more modular by moving creation and validation out of the monolithic
testPub method. The declarative philosophy of testPub is maintained.

This CL also includes several changes to the unit test library. Most notably, it
causes stack traces to be displayed for errors that are thrown during tests.

This was first submitted as r7566. That was rolled back due to a test error that
is fixed in this CL.

Review URL: https://chromiumcodereview.appspot.com//10382137

git-svn-id: http://dart.googlecode.com/svn/branches/bleeding_edge/dart@7573 260f80e4-7a28-3924-810f-c04153c831b5